Move generated request function for POST requests
This turns the "request" function that is present in all of the services
that uses a POST using queries or JSON into a function that is
centralized and called from every service.

Now things like building the host, or choosing the encoder and decoder
can be better controlled from one place.