Add REST and post request functions with basic procotol tests (#51)
This turns the "request" function that is present in all of the services
that uses a POST using queries or JSON into a function that is
centralized and called from every service.

It follows the same idea for REST requests.
It makes the modules for the services much easier to maintain and to
read, and also centralize the request code in one place.

Now things like building the host or choosing the encoder and decoder
can be better controlled from one place.

This commit also introduces protocol tests based in aws-sdk-go. The
protocol tests are important to ensure that modifications don't
break the lib.