Telemetry for AWS requests
This patch add Telemetry support for all AWS requests.

For context, what this patch aims is to eventually map those Telemetry
spans into [semantic OpenTelemetry traces][1]. This inform what kind of
metadata we need to provide on those spans:
* Client: which itself includes useful information like target region.
* ServiceMetadata: this is the most important one, with info about the
target Service itself;
* Action: the operation name, more on that bellow;
* Input: the input sent by the user, more on that bellow;

The `action` value is not provided on REST requests, and requires to
re-generate code. A companion PR can be sent to aws-codegen project, it
this feature is accepted. Given `AWS.Request` is a private implementation
for the generated code, I assume this change is not breaking from user
perspective.

The `input` AFAICT would be useful to extract service-specific
information like, for example, the table names of DynamoDB.

[1]: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/instrumentation/aws-sdk.md