Include a built-in Finch HTTP adapter
Following the approach of Swoosh, include a Finch-based adapter. The
patch keeps hackney adapter as default given it doesn't require any
extra setup.

In this patch the implementation of Hackney adapter got moved, which
should not be a breaking change for most cases - unless someone is
explicit referring to hackney implementation.

A small helper to update `http_client` key is also included, with the
intention to ease usage in cases where any create helper is used.

Example:

   AWS.Client.create()
   |> AWS.Client.put_http_client({AWS.HTTPClient.Finch, []})
   |> AWS.S3.list_buckets()

Using Application environment was considered as alternative to the
helper, but goes against what most of this library does.